PSORG Overhaul PR
This PR introduces restructuring of the MUIS, MAS, MLS, MPS, and DME startup sequences respectively

Numerous new packets have been added and introduced.

Added support for Pre-Medius 1.50 titles

Code organization has been done for cleaner reading.

+++ Server.Pipeline/Tcp/ScertDecoder.cs
@@ -64,7 +64,7 @@ namespace Server.Pipeline.Tcp
             var scertClient = context.GetAttribute(Constants.SCERT_CLIENT).Get();
 
             if (frameLength <= 0)
-                return BaseScertMessage.Instantiate((RT_MSG_TYPE)(id & 0x7F), null, new byte[0], scertClient.MediusVersion, scertClient.CipherService);
+                return BaseScertMessage.Instantiate((RT_MSG_TYPE)(id & 0x7F), null, new byte[0], (int)scertClient.MediusVersion, scertClient.CipherService);
 
             if (id >= 0x80)
             {
@@ -93,8 +93,7 @@ namespace Server.Pipeline.Tcp
 
             // 
             input.SetReaderIndex(input.ReaderIndex + totalLength + frameLengthInt);
-            return BaseScertMessage.Instantiate((RT_MSG_TYPE)id, hash, messageContents, scertClient.MediusVersion, scertClient.CipherService);
+            return BaseScertMessage.Instantiate((RT_MSG_TYPE)id, hash, messageContents, (int)scertClient.MediusVersion, scertClient.CipherService);
         }
-
     }
 }
